requirements.txt
click==6.7
Flask==0.12.2
Flask-Cache==0.13.1
Flask-SocketIO==3.0.0
imageio==2.3.0
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
MyQR==2.3.1
numpy==1.14.2
Pillow==5.1.0
PyMySQL==0.8.0
python-engineio==2.1.0
python-socketio==1.9.0
qrcode==6.0
six==1.11.0
Werkzeug==0.14.1
xlrd==1.1.0
